[refactor] 사용 않는 코드 제거 #47
